requirements.txt
supabase
python-dotenv
tinytuya
aiogram
pycoingecko
numpy
openai
cerebras_cloud_sdk
schedule
psutil